dms/requirements.txt
# Data Maintenance Service (DMS) Requirements
# Independent service for historical data maintenance, sync and backup

# Data fetching
yfinance>=0.2.0

# Database
influxdb>=5.3.0

# Web framework
flask>=3.0.0
flask-cors>=4.0.0
flask-login>=0.6.0
gunicorn>=21.2.0

# Utilities
pandas>=2.0.0
pyyaml>=6.0
requests>=2.31.0
python-dotenv>=1.0.0

# Scheduling
croniter>=2.0.0

# Other
python-dateutil>=2.8.0
